+++ src/main/kotlin/com/github/iammohdzaki/jsonbox/actions/OpenJsonQuickListAction.kt
@@ -1,6 +1,6 @@
 package com.github.iammohdzaki.jsonbox.actions
 
-import com.github.iammohdzaki.jsonbox.dialog.JsonBoxDialog
+import com.github.iammohdzaki.jsonbox.dialog.JsonBoxQuickDialog
 import com.github.iammohdzaki.jsonbox.utils.JsonBoxBundle
 import com.intellij.icons.AllIcons
 import com.intellij.openapi.actionSystem.ActionUpdateThread
@@ -8,21 +8,19 @@ import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 
 /**
- * Action to open the main JSON Parser dialog.
- * This dialog provides tools for parsing, formatting, and validating JSON content.
+ * Action to open the JSON Quick List dialog.
+ * This dialog allows users to manage and quickly access saved JSON snippets.
  */
-class OpenJsonParserDialogAction : AnAction() {
+class OpenJsonQuickListAction : AnAction() {
 
-    override fun actionPerformed(event: AnActionEvent) {
-        event.project?.let {
-            JsonBoxDialog(it, virtualFile = null).show()
-        }
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+        JsonBoxQuickDialog(project).show()
     }
 
     override fun update(e: AnActionEvent) {
         // Set text, description, and icon dynamically
-        e.presentation.text = JsonBoxBundle.message("jsonbox.title")
-        e.presentation.description = JsonBoxBundle.message("jsonbox.description")
+        e.presentation.text = JsonBoxBundle.message("jsonbox.quick.title")
         e.presentation.icon = AllIcons.FileTypes.Json
 
         // Optionally disable action if no project is open