+++ src/main/kotlin/com/github/iammohdzaki/jsonbox/utils/StringUtils.kt
@@ -1,9 +1,0 @@
-package com.github.iammohdzaki.jsonbox.utils
-
-object StringUtils {
-    const val PLUGIN_NAME = "JsonBox"
-    const val PLUGIN_DESCRIPTION = "Open a JSON file in JsonBox"
-    const val OPEN_IN_JSON_BOX = "Open in JsonBox"
-    const val NOTIFICATION_DESCRIPTION =
-        "You can open this JSON file in JsonBox for advanced formatting and validation."
-}