+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 # JsonBox Changelog
 
 ## [Unreleased]
-### Added
 
+### Added
 - Added Json Diff Feature to compare two JSON objects side by side
+- Added Json Quick List to save and manage JSON snippets
+- Added Notification support for important events
+- Moved to i18n for localization. Contributions are now welcomed.
 
 ### Fixed
 
-- Minor Fixes and Improvements
+- Improved JSON validation and error reporting
+- Minor Fixes and Performance Improvements